일정 기능 캘린더 수정 구현

+++ GroupwarePro/src/main/java/com/kh/spring/calendar/model/service/CalendarServiceImpl.java
@@ -15,15 +15,15 @@ import com.kh.spring.common.exception.CommException;
 @Service
 public class CalendarServiceImpl implements CalendarService {
 
-	@Autowired 
+	@Autowired
 	private SqlSessionTemplate sqlSession;
-	
+
 	@Autowired
 	private CalendarDao calendarDao;
-	
+
 	@Override
 	public ArrayList<Event> selectEventList() {
-		
+
 		ArrayList<Event> list = calendarDao.selectEventList(sqlSession);
 		return list;
 	}
@@ -45,30 +45,30 @@ public class CalendarServiceImpl implements CalendarService {
 
 		// insertEvent
 		int result = calendarDao.insertEvent(sqlSession, evt);
-		
+
 		if (result < 0) {
 			throw new CommException("이벤트 등록 실패");
 		}
-		
+
 		int evtNo = calendarDao.selectEvtNo(sqlSession) - 1; // 이게 마지막 번호가 아니라 다음 번호였음
-		
+
 		// insertAttendant
 		for (Attendant att : evt.getAttendantList()) {
 			att.setEvtNo(evtNo);
 			result = calendarDao.insertAttendant(sqlSession, att);
-			
+
 			if (result < 0) {
 				throw new CommException("참석자 등록 실패");
 			}
 		}
-		
+
 		// insertEvtCalReg
 		result = calendarDao.insertEvtCalReg(sqlSession, calNo, evtNo);
-		
+
 		if (result < 0) {
 			throw new CommException("캘린더 이벤트 등록 실패");
 		}
-		
+
 		return result;
 	}
 
@@ -76,21 +76,21 @@ public class CalendarServiceImpl implements CalendarService {
 	public int insertCalendar(Calendar cal) {
 		// insertCalendar
 		int result = calendarDao.insertCalendar(sqlSession, cal);
-		
+
 		if (result < 0) {
 			throw new CommException("캘린더 등록 실패");
 		}
-		
+
 		int calNo = calendarDao.selectCalNo(sqlSession) - 1;
 		cal.setCalNo(calNo);
-		
+
 		// insertCalReg
 		result = calendarDao.insertCalReg(sqlSession, cal);
-		
+
 		if (result < 0) {
 			throw new CommException("My 캘린더 등록 실패");
 		}
-		
+
 		return result;
 	}
 
@@ -101,7 +101,7 @@ public class CalendarServiceImpl implements CalendarService {
 
 	@Override
 	public int selectAttNo() {
-		return calendarDao.selectAttNo(sqlSession); 
+		return calendarDao.selectAttNo(sqlSession);
 	}
 
 	@Override
@@ -121,36 +121,68 @@ public class CalendarServiceImpl implements CalendarService {
 
 	@Override
 	public int updateEvent(Event evt) {
-		
+
 		ArrayList<Attendant> oldList = calendarDao.selectAttList(sqlSession, evt.getEvtNo());
-		
-		for (Attendant att : oldList) {
+
+		if (!isSameAtt(evt.getAttendantList(), oldList)) {
+
 			int result = calendarDao.deleteAttendant(sqlSession, evt.getEvtNo());
-			
+
 			if (result < 0) {
 				throw new CommException("참석자 삭제 실패");
 			}
-		}
-		
-		ArrayList<Attendant> newList = evt.getAttendantList();
-		
 
-		
+			// insertAttendant
+			for (Attendant att : evt.getAttendantList()) {
+				result = calendarDao.insertAttendant(sqlSession, att);
+
+				if (result < 0) {
+					throw new CommException("참석자 등록 실패");
+				}
+			}
+
+		}
+
 		return calendarDao.updateEvent(sqlSession, evt);
+	}
+
+	private boolean isSameAtt(ArrayList<Attendant> newList, ArrayList<Attendant> oldList) {
+
+		if (newList.size() != oldList.size()) {
+			return false;
+		} else {
+			for (int i = 0; i < newList.size(); i++) {
+				if (newList.get(i).getAttNo() != oldList.get(i).getAttNo()) {
+					return false;
+				}
+			}
+		}
+
+		return true;
 	}
 
 	@Override
 	public int updateCalendar(Calendar cal) {
+
+		int result = 0;
 		
 		if (cal.getColor() != null) {
-			int result = calendarDao.updateCalendarColor(cal);
-			
+			result = calendarDao.updateCalRegistration(sqlSession, cal);
+
 			if (result < 0) {
 				throw new CommException("캘린더 설정 수정 실패");
 			}
 		}
+
+		if (cal.getName() != null) {
+			result = calendarDao.updateCalRegistration(sqlSession, cal);
+
+			if (result < 0) {
+				throw new CommException("캘린더 수정 실패");
+			}			
+		}
 		
-		return calendarDao.updateCalendar(sqlSession, cal);
+		return result;
 	}
 
 }
